requirements.txt
numpy
pandas
requests
pytest
sqlalchemy
psycopg2
websocket-client